requirements.txt
streamlit>=1.31.1
google-generativeai>=0.5.0
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dotenv>=1.0.1
pypdf>=4.0.1
youtube-transcript-api>=0.6.2
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-docx>=1.1.0
fpdf2>=2.7.8
httpx>=0.27.0
Authlib>=1.3.0
